pyproject.toml
[tool.poetry]
name = "food-scun"
version = "0.1.0"
description = ""
authors = ["Ars <arsenzhanykulov@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
djangorestframework = "3.14"
django-cors-headers = "4.2,"
google-generativeai = "^0.8.6"
pillow = "10.1"
python-dotenv = "1.0"
psycopg2-binary = "^2.9.11"
django-allauth = "^65.14.0"
django-crispy-forms = "^2.5"
google-genai = "^1.61.0"
pyjwt = "^2.11.0"
djangorestframework-simplejwt = "^5.5.1"


[tool.poetry.group.dev.dependencies]
black = "^26.1.0"
isort = "^7.0.0"
autoflake = "^2.3.1"
flake8 = "^7.3.0"
mypy = "^1.19.1"
bandit = "^1.9.3"
pydocstyle = "^6.3.0"
pre-commit = "^4.5.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
